Moved EditingKeystrokeHandler from StandardEditor to Editor.
